refactor: contoller, service, domain, repository init
