+++ src/main/kotlin/io/hhplus/tdd/database/PointHistoryTable.kt
@@ -1,7 +1,7 @@
 package io.hhplus.tdd.database
 
-import io.hhplus.tdd.point.PointHistory
-import io.hhplus.tdd.point.TransactionType
+import io.hhplus.tdd.domain.PointHistory
+import io.hhplus.tdd.domain.TransactionType
 import org.springframework.stereotype.Component
 
 /**
