+++ src/main/kotlin/io/hhplus/tdd/domain/PointHistory.kt
@@ -1,4 +1,4 @@
-package io.hhplus.tdd.point
+package io.hhplus.tdd.domain
 
 data class PointHistory(
     val id: Long,
