+++ src/main/kotlin/io/hhplus/tdd/point/PointController.kt
@@ -1,5 +1,7 @@
 package io.hhplus.tdd.point
 
+import io.hhplus.tdd.domain.PointHistory
+import io.hhplus.tdd.domain.UserPoint
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 import org.springframework.web.bind.annotation.*
