+++ src/main/kotlin/io/hhplus/tdd/point/PointService.kt
@@ -1,0 +1,17 @@
+package io.hhplus.tdd.point
+
+import org.springframework.stereotype.Service
+
+@Service
+class PointService {
+
+    // 포인트 조회
+
+    // 포인트 충전/사용 내역 조회
+
+    // 포인트 충전
+
+    // 포인트 사용
+
+
+}
