+++ src/main/kotlin/io/hhplus/tdd/repository/UserPointRepository.kt
@@ -1,0 +1,8 @@
+package io.hhplus.tdd.repository
+
+import org.springframework.stereotype.Repository
+
+@Repository
+interface UserPointRepository {
+
+}
