+++ src/test/kotlin/io/hhplus/tdd/point/PointControllerTest.kt
@@ -1,0 +1,49 @@
+package io.hhplus.tdd.point
+
+import org.junit.jupiter.api.DisplayName
+import org.junit.jupiter.api.Test
+import org.springframework.boot.test.context.SpringBootTest
+
+@SpringBootTest
+class PointControllerTest {
+
+    @Test
+    @DisplayName("포인트 조회을 조회한다.")
+    fun point() {
+        // given
+
+        // when
+
+        // then
+    }
+
+    @Test
+    @DisplayName("포인트 충전/이용 내역을 조회한다.")
+    fun history() {
+        // given
+
+        // when
+
+        // then
+    }
+
+    @Test
+    @DisplayName("포인트를 충전한다.")
+    fun charge() {
+        // given
+
+        // when
+
+        // then
+    }
+
+    @Test
+    @DisplayName("포인트를 사용한다.")
+    fun use() {
+        // given
+
+        // when
+
+        // then
+    }
+}
