+++ src/test/kotlin/io/hhplus/tdd/point/PointServiceTest.kt
@@ -1,0 +1,5 @@
+package io.hhplus.tdd.point
+
+class PointServiceTest {
+
+}
