+++ src/test/kotlin/io/hhplus/tdd/point/UserPointTest.kt
@@ -1,0 +1,4 @@
+package io.hhplus.tdd.point
+
+class UserPointTest {
+}
